+++ test/gen_test.dart
@@ -1,4 +1,4 @@
-import 'package:path/path.dart' as p;
+import 'package:io/io.dart';
 import 'package:test/test.dart';
 
 import 'test_utils.dart';
@@ -10,22 +10,10 @@ void main() {
     final result = await testCommand(['gen']);
     final stdout = result.stdout;
 
-    // Since these projects have no code generation, the command should fail
-    expect(result.exitCode, isNot(0));
+    expect(result.exitCode, isNot(ExitCode.success.code));
 
-    // dart
     expectLine(stdout, ['dart_puby_test', 'dart $argString']);
-    // Explicit exclusion
-    expectLine(stdout, [p.join('dart_puby_test', 'example'), 'Skip']);
-
-    // flutter
     expectLine(stdout, ['flutter_puby_test', 'flutter $argString']);
-    // Explicit exclusion
-    expectLine(stdout, [p.join('flutter_puby_test', 'example'), 'Skip']);
-
-    // fvm
     expectLine(stdout, ['fvm_puby_test', 'fvm flutter $argString']);
-    // Explicit exclusion
-    expectLine(stdout, [p.join('fvm_puby_test', 'example'), 'Skip']);
   });
 }