+++ test/pub_test.dart
@@ -1,50 +1,95 @@
-import 'package:path/path.dart' as p;
+import 'package:io/io.dart';
+import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 
 import 'test_utils.dart';
 
 void main() {
-  test('[engine] pub get', () async {
-    final result = await testCommand(['get']);
-    final stdout = result.stdout;
+  group('[engine] pub get', () {
+    test('runs in all projects', () async {
+      final result = await testCommand(['get']);
+      final stdout = result.stdout;
 
-    expect(result.exitCode, 0);
+      expect(result.exitCode, ExitCode.success.code);
 
-    // project in build folder
-    expectLine(stdout, [p.join('build_folder_test', 'build', 'web'), 'Skip']);
+      expectLine(stdout, ['dart_puby_test', 'dart pub get']);
+      expectLine(stdout, ['flutter_puby_test', 'flutter pub get']);
+      expectLine(stdout, ['fvm_puby_test', 'fvm flutter pub get']);
+      expectLine(
+        stdout,
+        [path.join('fvm_puby_test', 'nested'), 'fvm flutter pub get'],
+      );
+    });
 
-    // dart
-    expectLine(stdout, ['dart_puby_test', 'dart pub get']);
-    expectLine(stdout, [p.join('dart_puby_test', 'example'), 'dart pub get']);
+    test('handles invlaid pubspec', () async {
+      final result = await testCommand(
+        ['get'],
+        projects: {
+          'invalid_pubspec_test': {
+            'pubspec.yaml': 'invalid',
+          },
+        },
+      );
+      final stdout = result.stdout;
 
-    // flutter
-    expectLine(stdout, ['flutter_puby_test', 'flutter pub get']);
-    // Default exclusion
-    expectLine(stdout, [p.join('flutter_puby_test', 'example'), 'Skip']);
-    // Flutter pub get should run in the example project anyways
-    expectLine(stdout, ['Resolving dependencies in ./example...']);
+      expect(result.exitCode, ExitCode.usage.code);
 
-    // fvm
-    expectLine(stdout, ['fvm_puby_test', 'fvm flutter pub get']);
-    // Default exclusion
-    expectLine(stdout, [p.join('fvm_puby_test', 'example'), 'Skip']);
-    // Flutter pub get should run in the example project anyways
-    // Can't test this with fvm since the output is the same as flutter
-    // expectLine(stdout, ['example', 'fvm flutter pub get']);
-    expectLine(
-      stdout,
-      [p.join('fvm_puby_test', 'nested'), 'fvm flutter pub get'],
-    );
+      expectLine(stdout, ['invalid_pubspec_test', 'Error parsing pubspec']);
+    });
 
-    // invalid_pubspec
-    expectLine(stdout, ['invalid_pubspec_test', 'Error parsing pubspec']);
+    group('excludes', () {
+      test('project in build folder', () async {
+        final result = await testCommand(
+          ['get'],
+          projects: {
+            'build_folder_test': {
+              'build/web/pubspec.yaml': pubspec('web'),
+            },
+          },
+        );
+        final stdout = result.stdout;
 
-    // transitive flutter
-    // This one should fail
-    // TODO: This isn't failing anymore for some reason. Remove this feature?
-    expectLine(stdout, ['transitive_flutter_test', 'dart pub get']);
+        expect(result.exitCode, ExitCode.success.code);
 
-    // This one should succeed
-    // expectLine(stdout, ['transitive_flutter_test', 'flutter pub get']);
+        expectLine(
+          stdout,
+          [path.join('build_folder_test', 'build', 'web'), 'Skip'],
+        );
+      });
+
+      group('example projects', () {
+        Future<void> skipsExample(
+          TestProjects projects, {
+          String match = 'Resolving dependencies in `./example`...',
+        }) async {
+          final result = await testCommand(['get'], projects: projects);
+          final stdout = result.stdout;
+
+          expect(result.exitCode, ExitCode.success.code);
+
+          expectLine(
+            stdout,
+            [path.join(projects.keys.first, 'example'), 'Skip'],
+          );
+          expectLine(stdout, [match]);
+        }
+
+        test('dart', () async {
+          await skipsExample(dartProject);
+        });
+
+        test('flutter', () async {
+          await skipsExample(flutterProject);
+        });
+
+        test('fvm', () async {
+          await skipsExample(
+            fvmProject,
+            // This is different because of the older Flutter version
+            match: 'Resolving dependencies in ./example...',
+          );
+        });
+      });
+    });
   });
 }