+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.36.0
+
+- Improves exit codes
+- Adds safety checks for FVM usage
+- Skips `pub get` in all example projects
+- Bumps Dart SDK constraint to `^3.0.0`
+
 ## 1.35.0
 - `puby link` no longer runs `pub get --offline` in flutter example projects
 - Fixes `--no-fvm` flag for `puby link`